+++ src/MetaheuristicsPlatform/Algorithms/Multiobjective/Spea2/Spea2References.cs
@@ -1,0 +1,12 @@
+using MetaheuristicsPlatform.Classification;
+namespace MetaheuristicsPlatform.Algorithms.Multiobjective.Spea2;
+public static class Spea2References
+{
+    public static ScientificReference ZitzlerLaumannsThiele2001 { get; } =
+        new(
+            "Zitzler, Laumanns & Thiele",
+            2001,
+            "SPEA2: Improving the Strength Pareto Evolutionary Algorithm",
+            "TIK Report 103, ETH Zurich",
+            "10.3929/ethz-a-004284029");
+}